+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
+- MV-223 - Doesn't access to a removed account
 - VIMEET-2011 - Hotfix: priorisation meeting request before planning
 ### Added
 - VIMEET-2030 - View checkin status in contacts list and agenda