+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
+- Improvment: Add cache to Crisp chat component
 - MV-223 - Doesn't access to a removed account
 - VIMEET-2011 - Hotfix: priorisation meeting request before planning
 ### Added